refactor(gir): Introduce callback types and restructure imports
- Added new `IntrospectedCallback` and `IntrospectedClassCallback` types in a dedicated `callback.ts` file to enhance organization and clarity.
- Updated import statements across multiple files to utilize the new callback types, improving maintainability.
- Removed redundant callback imports from various files, streamlining the codebase.
