+++ packages/lib/src/generators/json.ts
@@ -11,10 +11,9 @@ import {
     IntrospectedFunction,
     IntrospectedConstructor,
     IntrospectedFunctionParameter,
-    IntrospectedCallback,
-    IntrospectedDirectAllocationConstructor,
-    IntrospectedClassCallback
+    IntrospectedDirectAllocationConstructor
 } from "../gir/function.ts";
+import { IntrospectedCallback, IntrospectedClassCallback } from "../gir/callback.ts";
 import {
     IntrospectedClassFunction,
     IntrospectedStaticClassFunction,
