+++ packages/lib/src/gir/function.ts
@@ -28,7 +28,7 @@ import { GirDirection } from '@gi.ts/parser'
 
 import { IntrospectedNamespace } from "./namespace.ts";
 import { isIntrospectable } from "../utils/index.ts";
-import { getType, isInvalid, sanitizeMemberName, sanitizeIdentifierName, parseDoc, parseMetadata } from "../utils/index.ts";
+import { getType, isInvalid, sanitizeMemberName, sanitizeIdentifierName, parseDoc, parseMetadata, hasShadow } from "../utils/index.ts";
 import { IntrospectedClass } from "./class.ts";
 import { IntrospectedEnum } from "./enum.ts";
 import { IntrospectedSignal } from "./signal.ts";
@@ -39,9 +39,12 @@ import { IntrospectedBaseClass } from "./class.ts";
 
 import type { OptionsLoad } from "../types/index.ts";
 
-function hasShadow(obj: GirFunctionElement | GirMethodElement): obj is GirFunctionElement & { $: { shadows: string } } {
-    return obj.$["shadows"] != null;
-}
+// Import callback classes from the separate file
+import { IntrospectedCallback, IntrospectedClassCallback } from "./callback.ts";
+
+// ============================================================================
+// BASE FUNCTION CLASSES
+// ============================================================================
 
 export class IntrospectedFunction extends IntrospectedNamespaceMember {
     readonly parameters: IntrospectedFunctionParameter[];
@@ -322,6 +325,10 @@ export class IntrospectedFunction extends IntrospectedNamespaceMember {
     }
 }
 
+// ============================================================================
+// CONSTRUCTOR CLASSES
+// ============================================================================
+
 export class IntrospectedDirectAllocationConstructor extends IntrospectedClassMember {
     parameters: IntrospectedFunctionParameter[];
 
@@ -453,6 +460,10 @@ export class IntrospectedConstructor extends IntrospectedClassMember {
         return generator.generateConstructor(this) as ReturnType<T["generateConstructor"]>;
     }
 }
+
+// ============================================================================
+// FUNCTION PARAMETER
+// ============================================================================
 
 export class IntrospectedFunctionParameter extends IntrospectedBase<
     | IntrospectedClassFunction
@@ -635,6 +646,10 @@ export class IntrospectedFunctionParameter extends IntrospectedBase<
         return fp;
     }
 }
+
+// ============================================================================
+// CLASS FUNCTION CLASSES
+// ============================================================================
 
 export class IntrospectedClassFunction<
     Parent extends IntrospectedBaseClass | IntrospectedEnum = IntrospectedBaseClass | IntrospectedEnum
@@ -972,149 +987,4 @@ export class IntrospectedStaticClassFunction extends IntrospectedClassFunction {
     }
 }
 
-export class IntrospectedCallback extends IntrospectedFunction {
-    asFunctionType(): FunctionType {
-        return new FunctionType(
-            Object.fromEntries(this.parameters.map(p => [p.name, p.type] as const)),
-            this.return_type
-        );
-    }
 
-    copy({
-        parameters,
-        returnType,
-        outputParameters,
-        parent
-    }: {
-        parent?: GirModule;
-        parameters?: IntrospectedFunctionParameter[];
-        outputParameters?: IntrospectedFunctionParameter[];
-        returnType?: TypeExpression;
-    } = {}): IntrospectedCallback {
-        const cb = new IntrospectedCallback({
-            name: this.name,
-            raw_name: this.raw_name,
-            return_type: returnType ?? this.return_type,
-            parameters: parameters ?? this.parameters,
-            output_parameters: outputParameters ?? this.output_parameters,
-            namespace: parent ?? this.parent
-        })._copyBaseProperties(this);
-
-        cb.generics = [...this.generics];
-
-        return cb;
-    }
-
-    accept(visitor: GirVisitor): IntrospectedCallback {
-        const node = this.copy({
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            outputParameters: this.output_parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            returnType: visitor.visitType?.(this.return_type)
-        });
-
-        return visitor.visitCallback?.(node) ?? node;
-    }
-
-    static fromXML(element: GirCallbackElement, namespace: GirModule, options: OptionsLoad): IntrospectedCallback {
-        const ns = namespace;
-        const cb = IntrospectedFunction.fromXML(element, ns, options).asCallback();
-
-        const glibTypeName = element.$["glib:type-name"];
-        if (typeof glibTypeName === "string" && element.$["glib:type-name"]) {
-            cb.resolve_names.push(glibTypeName);
-
-            ns.registerResolveName(glibTypeName, ns.namespace, cb.name);
-        }
-
-        if (element.$["c:type"]) {
-            cb.resolve_names.push(element.$["c:type"]);
-
-            ns.registerResolveName(element.$["c:type"], ns.namespace, cb.name);
-        }
-
-        return cb;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateCallback"]> {
-        return generator.generateCallback(this) as ReturnType<T["generateCallback"]>;
-    }
-}
-
-export class IntrospectedClassCallback extends IntrospectedClassFunction {
-    asFunctionType(): FunctionType {
-        return new FunctionType(
-            Object.fromEntries(this.parameters.map(p => [p.name, p.type] as const)),
-            this.return_type
-        );
-    }
-
-    copy({
-        parameters,
-        returnType,
-        outputParameters,
-        parent
-    }: {
-        parent?: IntrospectedBaseClass;
-        parameters?: IntrospectedFunctionParameter[];
-        outputParameters?: IntrospectedFunctionParameter[];
-        returnType?: TypeExpression;
-    } = {}): IntrospectedClassCallback {
-        const cb = new IntrospectedClassCallback({
-            name: this.name,
-            return_type: returnType ?? this.return_type,
-            parameters: parameters ?? this.parameters,
-            output_parameters: outputParameters ?? this.output_parameters,
-            parent: parent ?? this.parent
-        })._copyBaseProperties(this);
-
-        cb.generics = [...this.generics];
-
-        return cb;
-    }
-
-    accept(visitor: GirVisitor): IntrospectedClassCallback {
-        const node = this.copy({
-            parameters: this.parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            outputParameters: this.output_parameters.map(p => {
-                return p.accept(visitor);
-            }),
-            returnType: visitor.visitType?.(this.return_type)
-        });
-
-        return visitor.visitClassCallback?.(node) ?? node;
-    }
-
-    static fromXML(
-        element: GirCallbackElement,
-        parent: IntrospectedBaseClass,
-        options: OptionsLoad
-    ): IntrospectedClassCallback {
-        const ns = parent.namespace;
-        const cb = IntrospectedClassFunction.fromXML(element, parent, options).asCallback();
-
-        const glibTypeName = element.$["glib:type-name"];
-        if (typeof glibTypeName === "string" && element.$["glib:type-name"]) {
-            cb.resolve_names.push(glibTypeName);
-
-            ns.registerResolveName(glibTypeName, ns.namespace, cb.name);
-        }
-
-        if (element.$["c:type"]) {
-            cb.resolve_names.push(element.$["c:type"]);
-
-            ns.registerResolveName(element.$["c:type"], ns.namespace, cb.name);
-        }
-
-        return cb;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateClassCallback"]> {
-        return generator.generateClassCallback(this) as ReturnType<T["generateClassCallback"]>;
-    }
-}
