+++ packages/lib/src/gir/signal.ts
@@ -12,7 +12,8 @@ import { IntrospectedClassMember } from "./introspected-class-member.ts";
 import type { Options } from "../types/introspected.ts";
 import { isIntrospectable } from "../utils/index.ts";
 import { IntrospectedClass } from "./class.ts";
-import { IntrospectedClassFunction, IntrospectedFunctionParameter, IntrospectedClassCallback } from "./function.ts";
+import { IntrospectedClassFunction, IntrospectedFunctionParameter } from "./function.ts";
+import { IntrospectedClassCallback } from "./callback.ts";
 import type { GirSignalElement, GirCallableParamElement } from "@gi.ts/parser";
 import { GirDirection } from "@gi.ts/parser";
 import { getType, parseDoc, parseMetadata } from "../utils/index.ts";
