+++ packages/lib/src/index.ts
@@ -1,14 +1,11 @@
 export * from './types/index.ts'
 export * from './constants.ts'
 export * from './dependency-manager.ts'
-// TODO: Finish migrating the types for TSDoc support
-export * from './types/ts-doc.ts'
 export * from './types/ts-doc-tag.ts'
 export * from './gir-module.ts'
 export * from './library-version.ts'
 export * from './logger.ts'
 export * from './messages.ts'
-
 export * from './utils/index.ts'
 export * from './gir.ts'
 export * from './gir/index.ts'
